Show list of commits
Let’s now fetch the list of commits and show their summaries, with links
to the same page but with a ?sha=xxxxxx query parameter.

The regular expression there splits a commit message into the summary
and the body, using the topmost sequence of two consecutive newlines as
the separator; if there is none, considers the whole message to be the
summary. The `[^\0]` pattern matches any character including newline but
not including the null character (though hopefully nobody in their right
mind uses null characters in commit messages).