Render markdown
GitHub has a Markdown API but it suffers from a fundamental flaw: It
takes a single Markdown string, returns an HTML string, and is
Rate-limited. This has two consequences:

* We would make a lot of requests, which would be fulfilled
asynchronously and cause much redisplay.
* The default rate limit for unauthenticated requests is 60 requests per
hour. By browsing the list of commits of a mature repository, we can
exhaust that instantly.

So let’s do markdown rendering on the client, using the [marked][]
library. It is distributed under the MIT license.

[marked]: https://github.com/chjj/marked